+++ src/controller/medico.py
@@ -1,11 +1,19 @@
+from fastapi import Depends, HTTPException
+from fastapi.security import OAuth2PasswordBearer
 from sqlalchemy.orm import Session
+from src.auth.jwt_handler import verify_token
 
 from src.entities.medico import Medico as medico
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")
 
 """ A partir de aqui hacemos metodos para los medicos"""
 
 
-def create_medico(db: Session, medico: medico):
+def create_medico(db: Session, medico: medico, token: str = Depends(oauth2_scheme)):
+    token_data = verify_token(token)
+    if token_data["rol"] not in ["admin"]:
+        raise HTTPException(status_code=403, detail="Forbidden")
     new_medico = medico(
         idMedico=str(medico.idMedico),
         especializacion=medico.especializacion,
@@ -18,15 +26,26 @@ def create_medico(db: Session, medico: medico):
     return new_medico
 
 
-def get_medico(db: Session, medico_id: str):
+def get_medico(db: Session, medico_id: str, token: str = Depends(oauth2_scheme)):
+    token_data = verify_token(token)
+    if token_data["rol"] not in ["admin", "medico", "enfermera"]:
+        raise HTTPException(status_code=403, detail="Forbidden")
     return db.query(medico).filter(medico.idMedico == medico_id).first()
 
 
-def get_medicos(db: Session):
+def get_medicos(db: Session, token: str = Depends(oauth2_scheme)):
+    token_data = verify_token(token)
+    if token_data["rol"] not in ["admin", "medico", "enfermera"]:
+        raise HTTPException(status_code=403, detail="Forbidden")
     return db.query(medico).all()
 
 
-def update_medico(db: Session, medico_id: str, medico: medico):
+def update_medico(
+    db: Session, medico_id: str, medico: medico, token: str = Depends(oauth2_scheme)
+):
+    token_data = verify_token(token)
+    if token_data["rol"] not in ["admin"]:
+        raise HTTPException(status_code=403, detail="Forbidden")
     db_medico = db.query(medico).filter(medico.idMedico == medico_id).first()
     if db_medico:
         db_medico.especializacion = medico.especializacion
@@ -37,7 +56,10 @@ def update_medico(db: Session, medico_id: str, medico: medico):
     return db_medico
 
 
-def delete_medico(db: Session, medico_id: str):
+def delete_medico(db: Session, medico_id: str, token: str = Depends(oauth2_scheme)):
+    token_data = verify_token(token)
+    if token_data["rol"] != "admin":
+        raise HTTPException(status_code=403, detail="Forbidden")
     db_medico = db.query(medico).filter(medico.idMedico == medico_id).first()
     if db_medico:
         db.delete(db_medico)
